[Feat] #28 - Rename method.py to enum.py and introduce FormEncoding class for enhanced form handling
